Add mypy

+++ app/services/scheduled_transactions.py
@@ -1,10 +1,11 @@
 from datetime import datetime
+from typing import Optional
 
 from app import models
 from app import repository as repo
 from app import schemas
 from app.logger import get_logger
-from app.utils.account_utils import has_user_access_to_account
+from app.services.accounts import AccountService
 from app.utils.exceptions import AccessDeniedError
 
 logger = get_logger(__name__)
@@ -12,7 +13,7 @@ logger = get_logger(__name__)
 
 async def get_transaction_list(
     user: models.User, account_id: int, date_start: datetime, date_end: datetime
-):
+) -> Optional[list[models.TransactionScheduled]]:
     """
     Retrieves a list of transactions within a specified period for a given account.
 
@@ -23,20 +24,28 @@ async def get_transaction_list(
         date_end: The end date of the period.
 
     Returns:
-        List[Transaction]: A list of transactions within the specified period.
+        list[Transaction]: A list of transactions within the specified period.
 
     Raises:
         None
     """
 
     account = await repo.get(models.Account, account_id)
+
+    if account is None:
+        return None
+
     if account.user_id == user.id:
         return await repo.get_scheduled_transactions_from_period(
             account_id, date_start, date_end
         )
 
+    return None
 
-async def get_transaction(user: models.User, transaction_id: int) -> models.Transaction:
+
+async def get_transaction(
+    user: models.User, transaction_id: int
+) -> Optional[models.TransactionScheduled]:
     """
     Retrieves a transaction by ID.
 
@@ -54,18 +63,20 @@ async def get_transaction(user: models.User, transaction_id: int) -> models.Tran
     transaction = await repo.get(models.TransactionScheduled, transaction_id)
 
     if transaction is None:
-        return
+        return None
 
     account = await repo.get(models.Account, transaction.account_id)
 
-    if account.user_id == user.id:
-        return transaction
+    if account is None:
+        return None
+
+    return transaction if account.user_id == user.id else None
 
 
 async def create_scheduled_transaction(
     user: models.User,
     transaction_information: schemas.ScheduledTransactionInformationCreate,
-) -> models.TransactionScheduled:
+) -> Optional[models.TransactionScheduled]:
     """
     Creates a scheduled transaction.
 
@@ -82,7 +93,10 @@ async def create_scheduled_transaction(
 
     account = await repo.get(models.Account, transaction_information.account_id)
 
-    if not has_user_access_to_account(user, account):
+    if account is None:
+        return None
+
+    if not AccountService.has_user_access_to_account(user, account):
         return None
 
     offset_account_id = transaction_information.offset_account_id
@@ -93,7 +107,7 @@ async def create_scheduled_transaction(
         if offset_account is None:
             return None
 
-        if not has_user_access_to_account(user, offset_account):
+        if not AccountService.has_user_access_to_account(user, offset_account):
             raise AccessDeniedError(
                 (
                     f"User[id: {user.id}] not allowed to access "
@@ -122,7 +136,7 @@ async def create_scheduled_transaction(
 
 async def delete_scheduled_transaction(
     current_user: models.User, transaction_id: int
-) -> bool:
+) -> Optional[bool]:
     """
     Deletes a scheduled transaction.
 
@@ -148,11 +162,11 @@ async def delete_scheduled_transaction(
 
     if transaction is None:
         logger.warning("Scheduled Transaction with ID %s not found.", transaction_id)
-        return
+        return None
 
     account = await repo.get(models.Account, transaction.account_id)
-    if current_user.id != account.user_id:
-        return
+    if account is None or current_user.id != account.user_id:
+        return None
 
     await repo.delete(transaction)
