+++ app/services/transactions.py
@@ -1,11 +1,11 @@
 from datetime import datetime
-from typing import List
+from typing import Optional
 
 from app import models
 from app import repository as repo
 from app import schemas
 from app.logger import get_logger
-from app.utils.account_utils import has_user_access_to_account
+from app.services.accounts import AccountService
 from app.utils.exceptions import AccessDeniedError
 from app.utils.log_messages import ACCOUNT_USER_ID_MISMATCH
 
@@ -27,7 +27,7 @@ class TransactionService:
         account_id: int,
         date_start: datetime,
         date_end: datetime,
-    ) -> List[models.Transaction]:
+    ) -> Optional[list[models.Transaction]]:
         """
         Retrieves a list of transactions within a specified period for a given account.
 
@@ -38,7 +38,7 @@ class TransactionService:
             date_end: The end date of the period.
 
         Returns:
-            List[Transaction]: A list of transactions within the specified period.
+            list[Transaction]: A list of transactions within the specified period.
 
         Raises:
             None
@@ -50,6 +50,10 @@ class TransactionService:
             account_id,
         )
         account = await repo.get(models.Account, account_id)
+
+        if account is None:
+            return None
+
         if account.user_id == user.id:
             logger.info("User ID verified. Retrieving transactions.")
             return await repo.get_transactions_from_period(
@@ -57,10 +61,11 @@ class TransactionService:
             )
 
         logger.warning(ACCOUNT_USER_ID_MISMATCH)
+        return None
 
     async def get_transaction(
         self, user: models.User, transaction_id: int
-    ) -> models.Transaction:
+    ) -> Optional[models.Transaction]:
         """
         Retrieves a transaction by ID.
 
@@ -82,21 +87,25 @@ class TransactionService:
 
         if transaction is None:
             logger.warning("Transaction with ID %s not found.", transaction_id)
-            return
+            return None
 
         account = await repo.get(models.Account, transaction.account_id)
+
+        if account is None:
+            return None
 
         if account.user_id == user.id:
             logger.info("User ID verified. Returning transaction.")
             return transaction
 
         logger.warning(ACCOUNT_USER_ID_MISMATCH)
+        return None
 
     async def create_transaction(
         self,
         user: models.User,
         transaction_information: schemas.TransactionInformationCreate,
-    ) -> models.Transaction:
+    ) -> Optional[models.Transaction]:
         """
         Creates a new transaction.
 
@@ -114,7 +123,10 @@ class TransactionService:
         logger.info("Creating new transaction for user %s", user.id)
         account = await repo.get(models.Account, transaction_information.account_id)
 
-        if not has_user_access_to_account(user, account):
+        if account is None:
+            return None
+
+        if not AccountService.has_user_access_to_account(user, account):
             logger.warning(ACCOUNT_USER_ID_MISMATCH)
             return None
 
@@ -160,7 +172,7 @@ class TransactionService:
         self,
         user: models.User,
         transaction_information: schemas.TransactionInformationCreate,
-    ) -> models.Transaction:
+    ) -> Optional[models.Transaction]:
         """
         Handles an offset transaction for a user.
 
@@ -177,13 +189,22 @@ class TransactionService:
 
         logger.info("Handling offset transaction for user %s", user.id)
         offset_account_id = transaction_information.offset_account_id
+
+        if offset_account_id is None:
+            return None
+
         offset_account = await repo.get(models.Account, offset_account_id)
+
+        if offset_account is None:
+            return None
 
         if user.id != offset_account.user_id:
             logger.warning("User ID does not match the offset account's User ID.")
             return None
 
-        transaction_information.amount = transaction_information.amount * -1
+        transaction_information.amount = schemas.RoundField(
+            transaction_information.amount * -1
+        )
         offset_account.balance += transaction_information.amount
 
         db_offset_transaction_information = models.TransactionInformation()
@@ -204,7 +225,7 @@ class TransactionService:
         current_user: models.User,
         transaction_id: int,
         transaction_information: schemas.TransactionInformtionUpdate,
-    ):
+    ) -> Optional[models.Transaction]:
         """
         Updates a transaction.
 
@@ -225,15 +246,19 @@ class TransactionService:
             transaction_id,
             current_user.id,
         )
-        transaction = await repo.get(models.Transaction, transaction_id)
+        transaction = await repo.get(
+            models.Transaction,
+            transaction_id,
+            load_relationships_list=[models.Transaction.account],
+        )
         if transaction is None:
-            logger.warning("Transaction with ID %s not found.", transaction_id)
-            return
+            return None
 
-        account = await repo.get(models.Account, transaction.account_id)
+        account = transaction.account
+
         if current_user.id != account.user_id:
             logger.warning(ACCOUNT_USER_ID_MISMATCH)
-            return
+            return None
 
         amount_updated = (
             round(transaction_information.amount, 2) - transaction.information.amount
@@ -242,21 +267,22 @@ class TransactionService:
         if transaction.offset_transactions_id:
             logger.info("Handling offset transaction for update.")
             offset_transaction = await repo.get(
-                models.Transaction, transaction.offset_transactions_id
+                models.Transaction,
+                transaction.offset_transactions_id,
+                load_relationships_list=[models.Transaction.account],
             )
-            offset_account = await repo.get(
-                models.Account, offset_transaction.account_id
-            )
+            if offset_transaction is None:
+                return None
 
-            if offset_account.user_id != current_user.id:
-                logger.warning("User ID does not match the offset account's User ID.")
-                return
+            offset_account = offset_transaction.account
+
+            if offset_account is None or offset_account.user_id != current_user.id:
+                return None
 
             offset_account.balance -= amount_updated
             offset_transaction.information.amount = transaction_information.amount * -1
 
         account_values = {"balance": account.balance + amount_updated}
-
         await repo.update(models.Account, account.id, **account_values)
 
         transaction_values = {
@@ -265,7 +291,6 @@ class TransactionService:
             "date": transaction_information.date,
             "category_id": transaction_information.category_id,
         }
-
         await repo.update(
             models.TransactionInformation,
             transaction.information.id,
@@ -276,7 +301,7 @@ class TransactionService:
 
     async def delete_transaction(
         self, current_user: models.User, transaction_id: int
-    ) -> bool:
+    ) -> Optional[bool]:
         """
         Deletes a transaction.
 
@@ -301,14 +326,17 @@ class TransactionService:
 
         if transaction is None:
             logger.warning("Transaction with ID %s not found.", transaction_id)
-            return
+            return None
 
         account = await repo.get(models.Account, transaction.account_id)
+
+        if account is None:
+            return None
+
         if current_user.id != account.user_id:
-            return
+            return None
 
         amount = transaction.information.amount
-        account.balance -= amount
 
         if transaction.offset_transaction:
             logger.info("Handling offset transaction for delete.")
@@ -316,9 +344,14 @@ class TransactionService:
             offset_account = await repo.get(
                 models.Account, offset_transaction.account_id
             )
+
+            if offset_account is None:
+                return None
+
             offset_account.balance += amount
             await repo.delete(transaction.offset_transaction)
 
+        account.balance -= amount
         await repo.delete(transaction)
 
         return True
